requirements.txt
praw==7.7.1
tweepy==4.14.0
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0
